+++ src/main/java/jsh/tftguide/recommend/application/RecommendService.java
@@ -5,5 +5,5 @@ import jsh.tftguide.champion.domain.Champion;
 import java.util.List;
 
 public interface RecommendService {
-    List<Champion> getBestChampions(List<Champion> champions, long level);
+    List<Champion> getBestChampions(List<Champion> champions, int level);
 }
